+++ go_bridge/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.2.0
+
+- getDynamicLibrary function probes for:
+    - current working directory
+    - bin subdirectory
+    - lib subdirectory
+    - go subdirectory
+
+## 1.1.0
+
+- Added getDynamicLibrary function.
+
 ## 1.0.0
 
 - Initial version.